requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
web3==7.2.0
pydantic==2.9.2